Made the program work on Windows again

+++ Sources/Logger.cpp
@@ -80,7 +80,7 @@ void Logger::Log(const std::string& aText, Level aLevel)
                   << aLocation.file_name() << ":" << aLocation.line() << ":" << aText;
 #else
         lLogEntry << std::put_time(std::gmtime(&lTimeAsTimeT), "%H:%M:%S:") << std::setfill('0') << std::setw(3)
-                  << lTimeMs.count() << ": " << mLogLevelTexts.at(aLevel) << ":" << aText;
+                  << lTimeMs.count() << ": " << mLogLevelTexts.at(static_cast<unsigned long>(aLevel)) << ":" << aText;
 #endif
 
         std::cout << lLogEntry.str() << std::endl;
